requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
langchain-core
